refactor(interpreter): split activation registration and validation
