+++ src/interpreter/src/activation/arms.rs
@@ -1,0 +1,12 @@
+use crate::{CompiledPattern, ValueKind};
+use super::ActivationPatternCapture;
+
+#[derive(Clone)]
+pub(super) struct PreflightActivationArm {
+    pub(super) pattern: CompiledPattern,
+    pub(super) captures: Vec<ActivationPatternCapture>,
+}
+pub(super) struct PreflightPatternedActivation {
+    pub(super) trigger_kind: ValueKind,
+    pub(super) arms: Vec<PreflightActivationArm>,
+}
