+++ src/interpreter/src/activation/mod.rs
@@ -4,64 +4,25 @@
 use crate::*;
 use std::collections::HashSet;
 
-macro_rules! activation_error {
-    ($n:ident,$m:expr) => {
-        #[derive(Debug, Clone)]
-        pub(crate) struct $n;
-        impl MechErrorKind for $n {
-            fn name(&self) -> &str {
-                stringify!($n)
-            }
-            fn message(&self) -> String {
-                $m.into()
-            }
-        }
-    };
-}
-activation_error!(
-    ActivationPatternCaptureKindUnsupported,
-    "The capture kind cannot be inferred from the activation trigger."
-);
-activation_error!(
-    ActivationPatternArmsNonExhaustive,
-    "Patterned activations require a final unguarded irrefutable arm."
-);
-activation_error!(
-    ActivationPatternWildcardMustBeLast,
-    "An unguarded wildcard activation arm must be last."
-);
-activation_error!(
-    ActivationPatternGuardMustBePure,
-    "Patterned activation guards must elaborate to a static pure expression graph."
-);
-activation_error!(
-    ActivationPatternGuardDependencyInvariant,
-    "The activation guard graph could not be attached to its match pulse."
-);
-activation_error!(
-    ActivationPatternBodyDependencyInvariant,
-    "The activation arm body could not sample its committed captures."
-);
-activation_error!(
-    ActivationPatternRegisterWriteUnsupported,
-    "Patterned activation register writes must target a whole local register."
-);
-activation_error!(
-    ActivationScopeTriggerWriteUnsupported,
-    "An activation scope cannot assign to its own trigger."
-);
-activation_error!(
-    ActivationPatternContextEffectUnsupported,
-    "Patterned activation context effects are not supported."
-);
-activation_error!(
-    ActivationPatternTriggerInvariant,
-    "Activation trigger root cells disagree with the resolved trigger."
-);
-activation_error!(
-    ActivationPatternTransactionBoolStateUnsupported,
-    "Patterned activation transaction state requires boolean values."
-);
+mod arms;
+mod errors;
+mod registration;
+mod validation;
+
+use arms::{PreflightActivationArm, PreflightPatternedActivation};
+pub(crate) use errors::{
+    ActivationPatternArmsNonExhaustive, ActivationPatternBodyDependencyInvariant,
+    ActivationPatternCaptureKindUnsupported, ActivationPatternContextEffectUnsupported,
+    ActivationPatternDefinitionUnsupported, ActivationPatternGuardDependencyInvariant,
+    ActivationPatternGuardMustBePure, ActivationPatternRegisterWriteUnsupported,
+    ActivationPatternTransactionBoolStateUnsupported, ActivationPatternTriggerInvariant,
+    ActivationPatternWildcardMustBeLast, ActivationScopeTriggerWriteUnsupported,
+};
+pub(crate) use registration::{activation_scope_entry_cells, elaborate_patterned_activation};
+use validation::{
+    preflight_patterned_activation, validate_patterned_arm_body,
+    validate_patterned_guard_expression,
+};
 
 fn transaction_bool_state(value: &Ref<bool>) -> MResult<Value> {
     #[cfg(any(feature = "bool", feature = "variable_define"))]
@@ -85,26 +46,6 @@ struct ActivationPatternCapture {
     kind: ValueKind,
     proposed: Value,
     committed: Value,
-}
-#[derive(Clone)]
-struct PreflightActivationArm {
-    pattern: CompiledPattern,
-    captures: Vec<ActivationPatternCapture>,
-}
-struct PreflightPatternedActivation {
-    trigger_kind: ValueKind,
-    arms: Vec<PreflightActivationArm>,
-}
-#[derive(Debug, Clone)]
-pub(crate) struct ActivationPatternDefinitionUnsupported;
-impl MechErrorKind for ActivationPatternDefinitionUnsupported {
-    fn name(&self) -> &str {
-        "ActivationPatternDefinitionUnsupported"
-    }
-    fn message(&self) -> String {
-        "This definition or declaration is not supported inside a patterned activation arm."
-            .to_string()
-    }
 }
 fn detached(v: &Value) -> Value {
     match v {
@@ -1041,732 +982,10 @@ interpreter_only!(Select);
 #[cfg(feature = "compiler")]
 interpreter_only!(Gate);
 
-fn pattern_is_irrefutable(pattern: &CompiledPattern, trigger_kind: &ValueKind) -> bool {
-    fn check(
-        pattern: &CompiledPattern,
-        trigger_kind: &ValueKind,
-        bindings: &mut HashSet<usize>,
-    ) -> bool {
-        match pattern {
-            CompiledPattern::Wildcard => true,
-            CompiledPattern::Binding { binding_index, .. } => {
-                bindings.insert(*binding_index)
-            }
-            CompiledPattern::ExpressionValue { .. }
-            | CompiledPattern::EnumVariant { .. }
-            | CompiledPattern::AtomTuple { .. } => false,
-            CompiledPattern::Tuple { elements } => {
-                let ValueKind::Tuple(kinds) = trigger_kind.deref_kind() else {
-                    return false;
-                };
-                elements.len() == kinds.len()
-                    && elements
-                        .iter()
-                        .zip(&kinds)
-                        .all(|(element, kind)| check(element, kind, bindings))
-            }
-            CompiledPattern::Array {
-                prefix,
-                spread,
-                suffix,
-            } => {
-                let ValueKind::Matrix(element_kind, shape) = trigger_kind.deref_kind() else {
-                    return false;
-                };
-                let minimum_len = prefix.len() + suffix.len();
-                let known_len = (!shape.is_empty()).then(|| shape.iter().product::<usize>());
-                match (known_len, spread) {
-                    (Some(len), None) if len != minimum_len => return false,
-                    (Some(len), Some(_)) if len < minimum_len => return false,
-                    (None, None) => return false,
-                    (None, Some(_)) if minimum_len != 0 => return false,
-                    _ => {}
-                }
-                if !prefix
-                    .iter()
-                    .chain(suffix)
-                    .all(|element| check(element, &element_kind, bindings))
-                {
-                    return false;
-                }
-                spread
-                    .as_ref()
-                    .and_then(|spread| spread.binding.as_deref())
-                    .map_or(true, |binding| {
-                        let middle_shape = known_len
-                            .map(|len| vec![1, len - minimum_len])
-                            .unwrap_or_default();
-                        check(
-                            binding,
-                            &ValueKind::Matrix(element_kind, middle_shape),
-                            bindings,
-                        )
-                    })
-            }
-        }
-    }
-
-    check(pattern, trigger_kind, &mut HashSet::new())
-}
-
-fn preflight_patterned_activation(
-    scope: &ActivationScope,
-    arms: &[ActivationArm],
-    trigger: &Value,
-    trigger_cells: &[ReactiveCellId],
-    i: &Interpreter,
-) -> MResult<PreflightPatternedActivation> {
-    arms.last().ok_or_else(|| {
-        MechError::new(ActivationPatternArmsNonExhaustive, None).with_tokens(scope.tokens())
-    })?;
-    let trigger_id = match &scope.trigger {
-        Expression::Var(var) => var.name.hash(),
-        _ => {
-            return Err(
-                MechError::new(ActivationPatternTriggerInvariant, None)
-                    .with_tokens(scope.trigger.tokens()),
-            );
-        }
-    };
-    for arm in arms {
-        if let Some(guard) = &arm.guard {
-            validate_patterned_guard_expression(guard, i)?;
-        }
-        validate_patterned_arm_body(&arm.body, trigger_id, trigger_cells, i)?;
-    }
-    if trigger.reactive_root_cell_ids() != trigger_cells {
-        return Err(
-            MechError::new(ActivationPatternTriggerInvariant, None).with_tokens(scope.tokens())
-        );
-    }
-    let trigger_kind = trigger.kind().deref_kind();
-    let mut compiled = Vec::new();
-    for a in arms {
-        let pattern = compile_pattern(&a.pattern, Some(&trigger_kind), i)?;
-        let captures = pattern
-            .binding_specs()
-            .into_iter()
-            .map(|binding| {
-                let kind = binding.kind.ok_or_else(|| {
-                    MechError::new(ActivationPatternCaptureKindUnsupported, None)
-                        .with_tokens(a.pattern.tokens())
-                })?;
-                let proposed = create_capture_slot_for_kind(&kind, i)
-                    .map_err(|error| error.with_tokens(a.pattern.tokens()))?;
-                let committed = create_capture_slot_for_kind(&kind, i)
-                    .map_err(|error| error.with_tokens(a.pattern.tokens()))?;
-                Ok(ActivationPatternCapture {
-                    id: binding.id,
-                    name: binding.name,
-                    kind,
-                    proposed,
-                    committed,
-                })
-            })
-            .collect::<MResult<Vec<_>>>()?;
-        compiled.push(PreflightActivationArm { pattern, captures });
-    }
-    let last = arms.last().unwrap();
-    if last.guard.is_some()
-        || !pattern_is_irrefutable(&compiled.last().unwrap().pattern, &trigger_kind)
-    {
-        return Err(
-            MechError::new(ActivationPatternArmsNonExhaustive, None).with_tokens(scope.tokens())
-        );
-    }
-    if arms[..arms.len() - 1]
-        .iter()
-        .any(|arm| arm.guard.is_none() && matches!(arm.pattern, Pattern::Wildcard))
-    {
-        return Err(
-            MechError::new(ActivationPatternWildcardMustBeLast, None)
-                .with_tokens(scope.tokens()),
-        );
-    }
-    Ok(PreflightPatternedActivation {
-        trigger_kind,
-        arms: compiled,
-    })
-}
-
-fn validation_error(kind: impl MechErrorKind + 'static, tokens: Vec<Token>) -> MResult<()> {
-    Err(MechError::new(kind, None).with_tokens(tokens))
-}
-
-fn validate_patterned_arm_body(
-    body: &ActivationArmBody,
-    trigger_id: u64,
-    trigger_cells: &[ReactiveCellId],
-    interpreter: &Interpreter,
-) -> MResult<()> {
-    match body {
-        ActivationArmBody::Block(body) => {
-            for (code, _) in body {
-                validate_patterned_code(code, trigger_id, trigger_cells, interpreter)?;
-            }
-            Ok(())
-        }
-        ActivationArmBody::Expression(expression) => validate_patterned_expression(expression),
-    }
-}
-fn validate_patterned_code(
-    code: &MechCode,
-    trigger_id: u64,
-    trigger_cells: &[ReactiveCellId],
-    interpreter: &Interpreter,
-) -> MResult<()> {
-    match code {
-        MechCode::Comment(_) => Ok(()),
-        MechCode::Expression(expression) => validate_patterned_expression(expression),
-        MechCode::Statement(statement) => {
-            validate_patterned_statement(statement, trigger_id, trigger_cells, interpreter)
-        }
-        MechCode::ActivationScope(_)
-        | MechCode::FunctionDefine(_)
-        | MechCode::FsmSpecification(_)
-        | MechCode::FsmImplementation(_)
-        | MechCode::Import(_)
-        | MechCode::Error(_, _) => {
-            validation_error(ActivationPatternDefinitionUnsupported, code.tokens())
-        }
-    }
-}
-fn validate_patterned_register_write(
-    target: &SliceRef,
-    expression: &Expression,
-    trigger_id: u64,
-    trigger_cells: &[ReactiveCellId],
-    interpreter: &Interpreter,
-    tokens: Vec<Token>,
-) -> MResult<()> {
-    if target.context.is_some() {
-        return validation_error(ActivationPatternContextEffectUnsupported, tokens);
-    }
-    let target_id = target.name.hash();
-    let aliases_trigger = interpreter
-        .symbols()
-        .borrow()
-        .get(target_id)
-        .is_some_and(|value| {
-            value
-                .borrow()
-                .reactive_root_cell_ids()
-                .iter()
-                .any(|cell| trigger_cells.contains(cell))
-        });
-    if target_id == trigger_id || aliases_trigger {
-        return validation_error(ActivationScopeTriggerWriteUnsupported, tokens);
-    }
-    // Indexed assignment implementations still mutate eagerly and do not
-    // implement the reactive-register staging contract.
-    if target.subscript.is_some() {
-        return validation_error(ActivationPatternRegisterWriteUnsupported, tokens);
-    }
-    validate_patterned_expression(expression)
-}
-
-fn validate_patterned_statement(
-    statement: &Statement,
-    trigger_id: u64,
-    trigger_cells: &[ReactiveCellId],
-    interpreter: &Interpreter,
-) -> MResult<()> {
-    match statement {
-        Statement::VariableDefine(definition)
-            if !definition.mutable && definition.var.context.is_none() =>
-        {
-            validate_patterned_expression(&definition.expression)
-        }
-        Statement::VariableDefine(definition) if definition.var.context.is_some() => {
-            validation_error(
-                ActivationPatternContextEffectUnsupported,
-                statement.tokens(),
-            )
-        }
-        Statement::VariableDefine(_) => {
-            validation_error(ActivationPatternDefinitionUnsupported, statement.tokens())
-        }
-        Statement::VariableAssign(assignment) => validate_patterned_register_write(
-            &assignment.target,
-            &assignment.expression,
-            trigger_id,
-            trigger_cells,
-            interpreter,
-            statement.tokens(),
-        ),
-        Statement::OpAssign(assignment) => validate_patterned_register_write(
-            &assignment.target,
-            &assignment.expression,
-            trigger_id,
-            trigger_cells,
-            interpreter,
-            statement.tokens(),
-        ),
-        Statement::ContextSend(_) => validation_error(
-            ActivationPatternContextEffectUnsupported,
-            statement.tokens(),
-        ),
-        _ => validation_error(ActivationPatternDefinitionUnsupported, statement.tokens()),
-    }
-}
-fn validate_patterned_expression(expression: &Expression) -> MResult<()> {
-    match expression {
-        Expression::Literal(_) | Expression::Var(_) => Ok(()),
-        Expression::Slice(slice) => validate_patterned_slice(slice),
-        Expression::Formula(factor) => validate_patterned_factor(factor),
-        Expression::FunctionCall(call) => {
-            for (_, expression) in &call.args {
-                validate_patterned_expression(expression)?;
-            }
-            Ok(())
-        }
-        Expression::Match(matched) => {
-            validate_patterned_expression(&matched.source)?;
-            for arm in &matched.arms {
-                validate_patterned_pattern(&arm.pattern)?;
-                if let Some(guard) = &arm.guard {
-                    validate_patterned_expression(guard)?;
-                }
-                validate_patterned_expression(&arm.expression)?;
-            }
-            Ok(())
-        }
-        Expression::Range(range) => validate_patterned_range(range),
-        Expression::Structure(structure) => validate_patterned_structure(structure),
-        Expression::SetComprehension(comprehension) => {
-            validate_patterned_expression(&comprehension.expression)?;
-            for qualifier in &comprehension.qualifiers {
-                validate_patterned_qualifier(qualifier)?;
-            }
-            Ok(())
-        }
-        Expression::MatrixComprehension(comprehension) => {
-            validate_patterned_expression(&comprehension.expression)?;
-            for qualifier in &comprehension.qualifiers {
-                validate_patterned_qualifier(qualifier)?;
-            }
-            Ok(())
-        }
-        Expression::FsmPipe(_) => {
-            validation_error(ActivationPatternDefinitionUnsupported, expression.tokens())
-        }
-    }
-}
-
-fn validate_patterned_guard_expression(
-    expression: &Expression,
-    interpreter: &Interpreter,
-) -> MResult<()> {
-    validate_patterned_expression(expression)?;
-    if guard_expression_is_not_static_pure(
-        expression,
-        interpreter,
-        &mut HashSet::new(),
-    ) {
-        validation_error(ActivationPatternGuardMustBePure, expression.tokens())
-    } else {
-        Ok(())
-    }
-}
-
-fn guard_expression_is_not_static_pure(
-    expression: &Expression,
-    interpreter: &Interpreter,
-    visiting_functions: &mut HashSet<u64>,
-) -> bool {
-    match expression {
-        Expression::Literal(_) | Expression::Var(_) => false,
-        Expression::Slice(slice) => slice
-            .subscript
-            .iter()
-            .any(|subscript| {
-                guard_subscript_is_not_static_pure(
-                    subscript,
-                    interpreter,
-                    visiting_functions,
-                )
-            }),
-        Expression::Formula(factor) => {
-            guard_factor_is_not_static_pure(factor, interpreter, visiting_functions)
-        }
-        Expression::FunctionCall(call) => {
-            if call.args.iter().any(|(_, expression)| {
-                guard_expression_is_not_static_pure(
-                    expression,
-                    interpreter,
-                    visiting_functions,
-                )
-            }) {
-                return true;
-            }
-            let function_id = call.name.hash();
-            let functions = interpreter.functions();
-            let functions = functions.borrow();
-            let user_function = functions.user_functions.get(&function_id).cloned();
-            let has_precompiled_function = functions.functions.contains_key(&function_id);
-            let native_guard_safety = functions
-                .function_compilers
-                .get(&function_id)
-                .map(|compiler| compiler.guard_safety());
-            drop(functions);
-            let Some(user_function) = user_function else {
-                if has_precompiled_function {
-                    return true;
-                }
-                return match native_guard_safety {
-                    Some(GuardFunctionSafety::PureStatic) | None => false,
-                    Some(GuardFunctionSafety::Unsupported) => true,
-                };
-            };
-            if !visiting_functions.insert(function_id) {
-                return true;
-            }
-            let eager = match user_function.code.match_arms.as_slice() {
-                [arm] if matches!(arm.pattern, Pattern::Wildcard) => {
-                    guard_expression_is_not_static_pure(
-                        &arm.expression,
-                        interpreter,
-                        visiting_functions,
-                    )
-                }
-                _ => true,
-            };
-            visiting_functions.remove(&function_id);
-            eager
-        }
-        Expression::Match(_)
-        | Expression::SetComprehension(_)
-        | Expression::MatrixComprehension(_)
-        | Expression::FsmPipe(_) => true,
-        Expression::Range(range) => {
-            guard_range_is_not_static_pure(range, interpreter, visiting_functions)
-        }
-        Expression::Structure(structure) => {
-            guard_structure_is_not_static_pure(structure, interpreter, visiting_functions)
-        }
-    }
-}
-
-fn guard_factor_is_not_static_pure(
-    factor: &Factor,
-    interpreter: &Interpreter,
-    visiting_functions: &mut HashSet<u64>,
-) -> bool {
-    match factor {
-        Factor::Expression(expression) => guard_expression_is_not_static_pure(
-            expression,
-            interpreter,
-            visiting_functions,
-        ),
-        Factor::Negate(factor)
-        | Factor::Not(factor)
-        | Factor::Parenthetical(factor)
-        | Factor::Transpose(factor) => {
-            guard_factor_is_not_static_pure(factor, interpreter, visiting_functions)
-        }
-        Factor::Term(term) => {
-            guard_factor_is_not_static_pure(
-                &term.lhs,
-                interpreter,
-                visiting_functions,
-            ) || term.rhs.iter().any(|(_, factor)| {
-                guard_factor_is_not_static_pure(
-                    factor,
-                    interpreter,
-                    visiting_functions,
-                )
-            })
-        }
-    }
-}
-
-fn guard_range_is_not_static_pure(
-    range: &RangeExpression,
-    interpreter: &Interpreter,
-    visiting_functions: &mut HashSet<u64>,
-) -> bool {
-    guard_factor_is_not_static_pure(&range.start, interpreter, visiting_functions)
-        || range
-            .increment
-            .as_ref()
-            .map_or(false, |(_, increment)| {
-                guard_factor_is_not_static_pure(
-                    increment,
-                    interpreter,
-                    visiting_functions,
-                )
-            })
-        || guard_factor_is_not_static_pure(
-            &range.terminal,
-            interpreter,
-            visiting_functions,
-        )
-}
-
-fn guard_subscript_is_not_static_pure(
-    subscript: &Subscript,
-    interpreter: &Interpreter,
-    visiting_functions: &mut HashSet<u64>,
-) -> bool {
-    match subscript {
-        Subscript::Brace(subscripts) | Subscript::Bracket(subscripts) => subscripts
-            .iter()
-            .any(|subscript| {
-                guard_subscript_is_not_static_pure(
-                    subscript,
-                    interpreter,
-                    visiting_functions,
-                )
-            }),
-        Subscript::Formula(factor) => {
-            guard_factor_is_not_static_pure(factor, interpreter, visiting_functions)
-        }
-        Subscript::Range(range) => {
-            guard_range_is_not_static_pure(range, interpreter, visiting_functions)
-        }
-        Subscript::All | Subscript::Dot(_) | Subscript::DotInt(_) | Subscript::Swizzle(_) => false,
-    }
-}
-
-fn guard_structure_is_not_static_pure(
-    structure: &Structure,
-    interpreter: &Interpreter,
-    visiting_functions: &mut HashSet<u64>,
-) -> bool {
-    match structure {
-        Structure::Empty => false,
-        Structure::Map(map) => map.elements.iter().any(|mapping| {
-            guard_expression_is_not_static_pure(
-                &mapping.key,
-                interpreter,
-                visiting_functions,
-            ) || guard_expression_is_not_static_pure(
-                &mapping.value,
-                interpreter,
-                visiting_functions,
-            )
-        }),
-        Structure::Matrix(matrix) => matrix.rows.iter().any(|row| {
-            row.columns
-                .iter()
-                .any(|column| {
-                    guard_expression_is_not_static_pure(
-                        &column.element,
-                        interpreter,
-                        visiting_functions,
-                    )
-                })
-        }),
-        Structure::Record(record) => record
-            .bindings
-            .iter()
-            .any(|binding| {
-                guard_expression_is_not_static_pure(
-                    &binding.value,
-                    interpreter,
-                    visiting_functions,
-                )
-            }),
-        Structure::Set(set) => set
-            .elements
-            .iter()
-            .any(|expression| {
-                guard_expression_is_not_static_pure(
-                    expression,
-                    interpreter,
-                    visiting_functions,
-                )
-            }),
-        Structure::Table(table) => table.rows.iter().any(|row| {
-            row.columns
-                .iter()
-                .any(|column| {
-                    guard_expression_is_not_static_pure(
-                        &column.element,
-                        interpreter,
-                        visiting_functions,
-                    )
-                })
-        }),
-        Structure::Tuple(tuple) => tuple
-            .elements
-            .iter()
-            .any(|expression| {
-                guard_expression_is_not_static_pure(
-                    expression,
-                    interpreter,
-                    visiting_functions,
-                )
-            }),
-        Structure::TupleStruct(tuple) => {
-            guard_expression_is_not_static_pure(
-                &tuple.value,
-                interpreter,
-                visiting_functions,
-            )
-        }
-    }
-}
-fn validate_patterned_pattern(pattern: &Pattern) -> MResult<()> {
-    match pattern {
-        Pattern::Expression(expression) => validate_patterned_expression(expression),
-        Pattern::Tuple(tuple) => {
-            for pattern in &tuple.0 {
-                validate_patterned_pattern(pattern)?;
-            }
-            Ok(())
-        }
-        Pattern::TupleStruct(tuple) => {
-            for pattern in &tuple.patterns {
-                validate_patterned_pattern(pattern)?;
-            }
-            Ok(())
-        }
-        Pattern::Array(array) => {
-            for pattern in array.prefix.iter().chain(&array.suffix) {
-                validate_patterned_pattern(pattern)?;
-            }
-            if let Some(spread) = &array.spread {
-                if let Some(binding) = &spread.binding {
-                    validate_patterned_pattern(binding)?;
-                }
-            }
-            Ok(())
-        }
-        Pattern::Wildcard => Ok(()),
-    }
-}
-fn validate_patterned_factor(factor: &Factor) -> MResult<()> {
-    match factor {
-        Factor::Expression(expression) => validate_patterned_expression(expression),
-        Factor::Negate(factor)
-        | Factor::Not(factor)
-        | Factor::Parenthetical(factor)
-        | Factor::Transpose(factor) => validate_patterned_factor(factor),
-        Factor::Term(term) => {
-            validate_patterned_factor(&term.lhs)?;
-            for (_, factor) in &term.rhs {
-                validate_patterned_factor(factor)?;
-            }
-            Ok(())
-        }
-    }
-}
-fn validate_patterned_range(range: &RangeExpression) -> MResult<()> {
-    validate_patterned_factor(&range.start)?;
-    if let Some((_, increment)) = &range.increment {
-        validate_patterned_factor(increment)?;
-    }
-    validate_patterned_factor(&range.terminal)
-}
-fn validate_patterned_slice(slice: &Slice) -> MResult<()> {
-    for subscript in &slice.subscript {
-        validate_patterned_subscript(subscript)?;
-    }
-    Ok(())
-}
-fn validate_patterned_subscript(subscript: &Subscript) -> MResult<()> {
-    match subscript {
-        Subscript::Brace(subscripts) | Subscript::Bracket(subscripts) => {
-            for subscript in subscripts {
-                validate_patterned_subscript(subscript)?;
-            }
-            Ok(())
-        }
-        Subscript::Formula(factor) => validate_patterned_factor(factor),
-        Subscript::Range(range) => validate_patterned_range(range),
-        Subscript::All | Subscript::Dot(_) | Subscript::DotInt(_) | Subscript::Swizzle(_) => Ok(()),
-    }
-}
-fn validate_patterned_structure(structure: &Structure) -> MResult<()> {
-    match structure {
-        Structure::Empty => Ok(()),
-        Structure::Map(map) => {
-            for mapping in &map.elements {
-                validate_patterned_expression(&mapping.key)?;
-                validate_patterned_expression(&mapping.value)?;
-            }
-            Ok(())
-        }
-        Structure::Matrix(matrix) => {
-            for row in &matrix.rows {
-                for column in &row.columns {
-                    validate_patterned_expression(&column.element)?;
-                }
-            }
-            Ok(())
-        }
-        Structure::Record(record) => {
-            for binding in &record.bindings {
-                validate_patterned_expression(&binding.value)?;
-            }
-            Ok(())
-        }
-        Structure::Set(set) => {
-            for expression in &set.elements {
-                validate_patterned_expression(expression)?;
-            }
-            Ok(())
-        }
-        Structure::Table(table) => {
-            for row in &table.rows {
-                for column in &row.columns {
-                    validate_patterned_expression(&column.element)?;
-                }
-            }
-            Ok(())
-        }
-        Structure::Tuple(tuple) => {
-            for expression in &tuple.elements {
-                validate_patterned_expression(expression)?;
-            }
-            Ok(())
-        }
-        Structure::TupleStruct(tuple) => validate_patterned_expression(&tuple.value),
-    }
-}
-fn validate_patterned_qualifier(qualifier: &ComprehensionQualifier) -> MResult<()> {
-    match qualifier {
-        ComprehensionQualifier::Generator((pattern, expression)) => {
-            validate_patterned_pattern(pattern)?;
-            validate_patterned_expression(expression)
-        }
-        ComprehensionQualifier::Filter(expression) => validate_patterned_expression(expression),
-        ComprehensionQualifier::Let(definition) if definition.mutable => {
-            validation_error(ActivationPatternDefinitionUnsupported, definition.tokens())
-        }
-        ComprehensionQualifier::Let(definition) if definition.var.context.is_some() => {
-            validation_error(
-                ActivationPatternContextEffectUnsupported,
-                definition.tokens(),
-            )
-        }
-        ComprehensionQualifier::Let(definition) => {
-            validate_patterned_expression(&definition.expression)
-        }
-    }
-}
-
 struct ElaboratedPatternGuard {
     finalizer_node: ReactiveNodeId,
     node_start: usize,
     node_end: usize,
-}
-
-pub(crate) fn activation_scope_entry_cells(
-    interpreter: &Interpreter,
-) -> Vec<ReactiveCellId> {
-    let symbols = interpreter.symbols();
-    let symbols = symbols.borrow();
-    let mut cells = Vec::new();
-    for symbol in symbols.symbols.values() {
-        for cell in symbol.borrow().reactive_cell_ids() {
-            if !cells.contains(&cell) {
-                cells.push(cell);
-            }
-        }
-    }
-    cells
 }
 
 fn elaborate_patterned_arm_guard(
@@ -1866,280 +1085,6 @@ fn elaborate_patterned_arm_guard(
     }
     symbols.borrow_mut().restore(symbol_snapshot);
     result
-}
-
-fn elaborate_patterned_arm_body(
-    arm: &ActivationArm,
-    captures: &[ActivationPatternCapture],
-    pulse: &Value,
-    interpreter: &InterpreterExecution<'_>,
-) -> MResult<(usize, usize)> {
-    let symbols = interpreter.symbols();
-    let symbol_snapshot = symbols.borrow().snapshot();
-    let plan = interpreter.plan();
-    let original_scope_depth = plan.activation_registration_depth();
-    {
-        let mut symbols = symbols.borrow_mut();
-        for capture in captures {
-            symbols.mutable_variables.remove(&capture.id);
-            symbols.insert(capture.id, capture.committed.clone(), false);
-            symbols
-                .dictionary
-                .borrow_mut()
-                .insert(capture.id, capture.name.clone());
-        }
-    }
-    let body_node_start = plan.len();
-    plan.push_activation_registration_scope_with_sampled_cells(
-        pulse.reactive_root_cell_ids(),
-        activation_scope_entry_cells(interpreter),
-    );
-    let body_result = (|| -> MResult<()> {
-        match &arm.body {
-            ActivationArmBody::Block(body) => {
-                for (code, _) in body {
-                    crate::mech_code(code, interpreter)?;
-                }
-                Ok(())
-            }
-            ActivationArmBody::Expression(expression) => {
-                crate::expression(expression, None, interpreter)?;
-                Ok(())
-            }
-        }
-    })();
-    while plan.activation_registration_depth() > original_scope_depth {
-        plan.pop_activation_registration_scope();
-    }
-    symbols.borrow_mut().restore(symbol_snapshot);
-    body_result?;
-    let body_node_end = plan.len();
-    {
-        let mut plan = plan.borrow_mut();
-        for node in body_node_start..body_node_end {
-            for capture in captures {
-                let cell = capture.committed.reactive_root_cell_ids()[0];
-                if !plan.add_sampled_dependency(node, cell) {
-                    return Err(MechError::new(
-                        ActivationPatternBodyDependencyInvariant,
-                        None,
-                    ));
-                }
-            }
-        }
-    }
-    Ok((body_node_start, body_node_end))
-}
-
-fn elaborate_patterned_activation_inner(
-    arms: &[ActivationArm],
-    trigger: Value,
-    preflight: PreflightPatternedActivation,
-    i: &InterpreterExecution<'_>,
-) -> MResult<Value> {
-    if trigger.kind().deref_kind() != preflight.trigger_kind {
-        return Err(MechError::new(ActivationPatternTriggerInvariant, None));
-    }
-    let compiled = preflight.arms;
-    let plan = i.plan();
-    let _persistent_user_function_plan =
-        crate::functions::PersistentUserFunctionPlanScope::enter(i);
-    let pattern_expression_values = compiled
-        .iter()
-        .map(|arm| {
-            arm.pattern
-                .expressions()
-                .iter()
-                .map(|expression| crate::expression(expression, None, i))
-                .collect::<MResult<Vec<_>>>()
-        })
-        .collect::<MResult<Vec<_>>>()?;
-    drop(_persistent_user_function_plan);
-    // Seed proposal storage before guard graphs are elaborated. Composite
-    // guard expressions may need the current proposal shape to compile, but
-    // eligibility and selection are still determined by the runtime graph
-    // initialization below.
-    for (arm, expression_values) in compiled.iter().zip(&pattern_expression_values) {
-        let pattern_match =
-            match_compiled_pattern_with_values(&arm.pattern, &trigger, expression_values)?;
-        ReactiveBindingSink {
-            captures: &arm.captures,
-        }
-        .commit(&pattern_match)?;
-    }
-    let (scope_gen, scope_v) = generation();
-    let scope_node = plan
-        .borrow_mut()
-        .register(Box::new(ScopePulse { out: scope_gen }), &[trigger.clone()])?;
-    let (mut matcher_nodes, mut completions, mut matched) = (Vec::new(), Vec::new(), Vec::new());
-    for (arm, expression_values) in compiled.iter().zip(&pattern_expression_values) {
-        let (o, v) = generation();
-        let f = Ref::new(false);
-        let mut inputs = Vec::with_capacity(2 + expression_values.len());
-        inputs.push(scope_v.clone());
-        inputs.push(trigger.clone());
-        inputs.extend(expression_values.iter().cloned());
-        let n = plan.borrow_mut().register(
-            Box::new(Matcher {
-                pattern: arm.pattern.clone(),
-                trigger: trigger.clone(),
-                expression_values: expression_values.clone(),
-                captures: arm.captures.clone(),
-                matched: f.clone(),
-                out: o,
-            }),
-            &inputs,
-        )?;
-        matcher_nodes.push(n);
-        completions.push(v);
-        matched.push(f);
-    }
-    let (mut finalizers, mut guards, mut eligible, mut done) =
-        (Vec::new(), Vec::new(), Vec::new(), Vec::new());
-    for n in 0..arms.len() {
-        let e = Ref::new(false);
-        if let Some(guard) = &arms[n].guard {
-            let (match_gate_out, match_gate_pulse) = generation();
-            let match_gate_node = plan.borrow_mut().register(
-                Box::new(MatchGate {
-                    matched: matched[n].clone(),
-                    out: match_gate_out,
-                }),
-                &[completions[n].clone()],
-            )?;
-            let (unmatched_out, unmatched_done) = generation();
-            let unmatched_finalizer = plan.borrow_mut().register(
-                Box::new(UnmatchedFinalize {
-                    matched: matched[n].clone(),
-                    eligible: e.clone(),
-                    out: unmatched_out,
-                }),
-                &[completions[n].clone()],
-            )?;
-            let (guard_out, guard_done) = generation();
-            let elaborated = elaborate_patterned_arm_guard(
-                guard,
-                &compiled[n].captures,
-                &match_gate_pulse,
-                &e,
-                guard_out,
-                i,
-            )?;
-            finalizers.push(unmatched_finalizer);
-            guards.push(Some(PatternActivationGuardRegistration {
-                match_gate_node,
-                guard_finalizer_node: elaborated.finalizer_node,
-                guard_node_start: elaborated.node_start,
-                guard_node_end: elaborated.node_end,
-            }));
-            done.push(unmatched_done);
-            done.push(guard_done);
-        } else {
-            let (out, arm_done) = generation();
-            finalizers.push(plan.borrow_mut().register(
-                Box::new(Finalize {
-                    matched: matched[n].clone(),
-                    eligible: e.clone(),
-                    out,
-                }),
-                &[completions[n].clone()],
-            )?);
-            guards.push(None);
-            done.push(arm_done);
-        }
-        eligible.push(e);
-    }
-    let (o, selection) = generation();
-    let selected = Ref::new(usize::MAX);
-    let selector = plan.borrow_mut().register(
-        Box::new(Select {
-            eligible: eligible.clone(),
-            selected: selected.clone(),
-            out: o,
-        }),
-        &done,
-    )?;
-    let private_scope_cell = scope_v.reactive_root_cell_ids()[0];
-    plan.solve_dirty_cells(&[private_scope_cell])?;
-    let initially_selected = *selected.borrow();
-    if initially_selected >= compiled.len() {
-        return Err(MechError::new(ActivationPatternArmsNonExhaustive, None));
-    }
-    commit_proposed_captures(&compiled[initially_selected].captures)?;
-    let (mut gates, mut pulses) = (Vec::new(), Vec::new());
-    for arm in 0..arms.len() {
-        let (o, v) = generation();
-        gates.push(plan.borrow_mut().register(
-            Box::new(Gate {
-                arm,
-                selected: selected.clone(),
-                captures: compiled[arm].captures.clone(),
-                out: o,
-            }),
-            &[selection.clone()],
-        )?);
-        pulses.push(v);
-    }
-    let mut ranges = Vec::new();
-    for (arm, compiled_arm) in arms.iter().zip(&compiled) {
-        ranges.push(elaborate_patterned_arm_body(
-            arm,
-            &compiled_arm.captures,
-            &pulses[ranges.len()],
-            i,
-        )?);
-    }
-    let registration = PatternActivationRegistration {
-        scope_pulse_node: scope_node,
-        selector_node: selector,
-        arms: (0..arms.len())
-            .map(|n| PatternActivationArmRegistration {
-                matcher_node: matcher_nodes[n],
-                finalizer_node: finalizers[n],
-                guard: guards[n].clone(),
-                gate_node: gates[n],
-                pulse_cell: pulses[n].reactive_root_cell_ids()[0],
-                body_node_start: ranges[n].0,
-                body_node_end: ranges[n].1,
-                captures: compiled[n]
-                    .captures
-                    .iter()
-                    .map(|c| PatternActivationCaptureRegistration {
-                        id: c.id,
-                        kind: c.kind.clone(),
-                        cell: c.committed.reactive_root_cell_ids()[0],
-                    })
-                    .collect(),
-            })
-            .collect(),
-    };
-    plan.borrow_mut().register_pattern_activation(registration);
-    Ok(Value::Empty)
-}
-
-pub(crate) fn elaborate_patterned_activation(
-    scope: &ActivationScope,
-    arms: &[ActivationArm],
-    trigger: Value,
-    trigger_cells: Vec<ReactiveCellId>,
-    interpreter: &InterpreterExecution<'_>,
-) -> MResult<Value> {
-    let preflight =
-        preflight_patterned_activation(scope, arms, &trigger, &trigger_cells, interpreter)?;
-    let plan = interpreter.plan();
-    let checkpoint = plan.checkpoint();
-    let program_dictionary = interpreter.state.borrow().dictionary.clone();
-    let dictionary_snapshot = program_dictionary.borrow().clone();
-    match elaborate_patterned_activation_inner(arms, trigger, preflight, interpreter) {
-        Ok(value) => Ok(value),
-        Err(error) => {
-            *program_dictionary.borrow_mut() = dictionary_snapshot;
-            match plan.rollback(checkpoint) {
-                Ok(()) => Err(error),
-                Err(rollback_error) => Err(rollback_error),
-            }
-        }
-    }
 }
 
 #[cfg(test)]
